usePlacesAutocompleteUtils hook
